+++ packages/commonwealth/client/scripts/views/components/ThreadTokenDrawer/index.ts
@@ -1,2 +1,4 @@
 export { ThreadTokenDrawer } from './ThreadTokenDrawer';
-export { ThreadTokenDrawerTrigger } from './ThreadTokenDrawerTrigger';
+export { ThreadTokenDrawerTrigger } from './ThreadTokenDrawerTrigger/ThreadTokenDrawerTrigger';
+export { TokenHoldersTab } from './TokenHolderTab/TokenHoldersTab';
+export { TradeActivityTab } from './TradeActivityTab/TradeActivityTab';
